### 2026-08-18 #11056 @olexandr-konovalov
Hello @thraraujo @kaashif @timbode,

I am looking for one more reviewer for this JOSS paper. Would one of you be interested and able to review it, please?

If not, your suggestions for potential reviewers would be most appreciated!
