### 2026-08-18 #11056 @kaashif
Sure, I'd be happy to help!

On Tue, 18 Aug 2026, 21:26 Olexandr Konovalov, ***@***.***>
wrote:

> *olexandr-konovalov* left a comment (openjournals/joss-reviews#11056)
> <https://github.com/openjournals/joss-reviews/issues/11056#issuecomment-5333649217>
>
> Hello @thraraujo <https://github.com/thraraujo> @kaashif
> <https://github.com/kaashif> @timbode <https://github.com/timbode>,
>
> I am looking for one more reviewer for this JOSS paper. Would one of you
> be interested and able to review it, please?
>
> If not, your suggestions for potential reviewers would be most appreciated!
>
