### 2026-08-18 #11169 @editorialbot
[REVIEW]: GroupFunctions.jl: computing individual entries of the irreducible representations of the unitary group U(d)
**Submitting author:** @knszm (<a href="http://orcid.org/0000-0001-7676-1605">Konrad Szymański</a>)
**Repository:** https://github.com/davidamaro/GroupFunctions.jl
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** 0.3.4
