### 2026-08-18 #11056 @olexandr-konovalov
@kaashif @SoongNoonien wonderful - thanks to both of you for the speedy replies!

I have started the review at https://github.com/openjournals/joss-reviews/issues/11169 - please see it for further instructions, in particular how to generate your individual review checklists. Thank you!
