### 2026-08-22 af5a732
Address review feedback on conversion API
