+++ paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 
 # Summary
-[`GroupFunctions.jl`](https://github.com/davidamaro/GroupFunctions.jl)[^author-contributions] is a Julia library for computing individual matrix elements of irreducible representations of $\mathrm{U}(d)$. These matrix elements, called group functions, can be evaluated symbolically or numerically. For $\mathrm{SU}(2)$, they reduce to the Wigner $D$-functions. The library computes these matrix elements in a carrier-space basis enumerated by Gelfand-Tsetlin patterns [@GelfandTsetlin1950]. It can also compute entire representation operators, construct input unitaries from parameterisations common in quantum optics, translate Gelfand-Tsetlin patterns into occupation-number kets, and compute the associated Schur functions. Results can be exported in a form compatible with Mathematica.
+[`GroupFunctions.jl`](https://github.com/davidamaro/GroupFunctions.jl)[^author-contributions] is a Julia library for computing individual matrix elements of irreducible representations of $\mathrm{U}(d)$. These matrix elements, called group functions, can be evaluated symbolically or numerically. For $\mathrm{SU}(2)$, they reduce to the Wigner $D$-functions. The library computes these matrix elements in a carrier-space basis enumerated by Gelfand-Tsetlin patterns [@GelfandTsetlin1950]. It can also compute entire representation operators, construct input unitaries from parameterisations common in quantum optics, translate Gelfand-Tsetlin patterns into occupation-number kets, and compute the associated Schur functions.
 
 [^author-contributions]: David Amaro-Alcalá developed the package, its algorithms, and its test suite, and prepared the initial documentation. Konrad Szymański substantially revised and expanded the documentation and developed additional examples, and contributed to the standardisation of the API.
      
