### 2026-08-22 c9a9255
Stabilize symbolic matrix doctest
