### 2026-08-31 #11169 @davidamaro
> I have two comments so far:
>
> * There actually is a performance claim in the paper that is not explained: "Although related packages exist (see
>   below), to our knowledge, none are designed to compute the representation matrix symbolically one matrix element at a time – the cheaper route when a calculation requires only a few of them" - it is not clear to me from the text why this is true. Reading through the code there is a lot of computation of cosets, lots of matrices, etc, although I am sure with a detailed analysis it's clear what the trick is, some elaboration on why the method used here skips lots of computations needed for the full linear maps would be great.
>   If this were a minor part of a larger project it might be fine to leave this unexplained, but since it is the whole point of this package and the core of why this is a novel contribution, I think explaining the performance advantage here is important.
>   From an algorithmic perspective it would be interesting to know whether the shortcuts taken are actually asymptotic improvements in the algorithm or "just" large practical improvements.
> * The audience is a bit unclear - physicists? I agree with [@SoongNoonien](https://github.com/SoongNoonien)'s commments on matrix entry being clearer and notation possibly being clearer or more accessible to a wider audience. I am also from a mathematics background and unfamiliar with physics notation - perhaps including both notations would be best? If the audience is physicists then physics notation may be fine.

Hi @kaashif, I’m replying to the first part of your comment since it seems most urgent. Konrad and I will address the rest in a follow-up message.

I’ll address this part separately from your original comment because I think it will help clarify your concern.

> **(kaashif)** There actually is a performance claim in the paper that is not explained: "Although related packages exist (see below), to our knowledge, none are designed to compute the representation matrix symbolically one matrix element at a time – the cheaper route when a calculation requires only a few of them" - it is not clear to me from the text why this is true.

We recognise that our earlier wording was unclear. The phrase “a few of them” did not specify the number of entries or the relevant asymptotic regime. To address this, we now limit our claim to a single exact symbolic $D$-function and compare it only with the standard full-matrix Lie-algebra approach.

Specifically, we will delete the original sentence:

> “Although related packages exist (see below), to our knowledge, none are designed to compute the representation matrix symbolically one matrix element at a time—the cheaper route when a calculation requires only a few of them.”

We will replace it with:

> “Although related software exists (see below), to our knowledge none is designed to compute a single representation-matrix entry as an exact symbolic \(D\)-function. We therefore compare the cost of computing one such entry with the conventional route that constructs the Lie-algebra representation matrices and exponentiates the resulting full matrix.”

We will then add the following supporting paragraph:

Let $\lambda\vdash N$, with $N>1$, and let the compatible basis states $A,B$ be supplied directly. For fixed $\lambda,A,B$ and $d\to\infty$, the hook–content formula gives
$$D_\lambda(d)=\frac{f^\lambda}{N!}d^N+O(d^{N-1})=\Theta(d^N).$$
The present implementation evaluates one symbolic $D$-function using $\Theta(d^3)$ structural operations and $\Theta(d^2)$ peak storage. By contrast, even if construction of the $D_\lambda(d)\times D_\lambda(d)$ Lie-algebra matrix is granted at no cost, conventional unstructured scaling-and-squaring/Padé exponentiation with classical dense arithmetic requires $\Theta(D_\lambda(d)^3)=\Theta(d^{3N})$ operations and $\Theta(D_\lambda(d)^2)=\Theta(d^{2N})$ storage. Consequently, the direct-to-exponentiation ratios are
$$\Theta\!\left(d^{-3(N-1)}\right) \quad\text{and}\quad \Theta\!\left(d^{-2(N-1)}\right),$$
respectively. Thus, computing one requested $D$-function has an asymptotic advantage over constructing the complete representation matrix in this fixed-$N$, growing-$d$ regime.

> Reading through the code there is a lot of computation of cosets, lots of matrices, etc,

We agree with this assessment.

> although I am sure with a detailed analysis it's clear what the trick is, some elaboration on why the method used here skips lots of computations needed for the full linear maps would be great.

To be clear, there is no hidden trick in our approach. We use Grabmeier's formula directly. The main advantage over current methods is that we do not need to compute a matrix exponential. To support this, we compare the cost of our algorithm with that of matrix exponentiation, not counting the cost of building the Lie-algebra matrices.

> If this were a minor part of a larger project it might be fine to leave this unexplained, but since it is the whole point of this package and the core of why this is a novel contribution, I think explaining the performance advantage here is important.

We agree with this assessment. As noted earlier, the software is unique because it does not use the usual method for computing exponentials of Lie-algebra elements.

> From an algorithmic perspective it would be interesting to know whether the shortcuts taken are actually asymptotic improvements in the algorithm or "just" large practical improvements.

As explained earlier, our approach lowers the computational complexity compared to the basic method of exponentiating Lie-algebra elements.

We hope this helps clarify what we meant. As we mentioned earlier, we will respond to the rest of the comments about notation and terminology in a separate message.
