### 2026-09-01 #11169 @knszm
Let me reply to the more notation-oriented comments of the reviewers.

@SoongNoonien raised the issue of lack of clarity of the terms "irrep" and "matrix element", as well as the bra-ket notation we use in the paper and the documentation:

> - First of all, the term "irrep" is, in my opinion, totally fine for verbal communication or blackboard writing but in the paper and the documentation it is used without stating that it is a abbreviation of "irreducible representation" (…)
> - The other term which is unclear to me is "matrix element". (…)
> - (…) My suggestion would be something along the lines of $out^T \cdot D^{(\lambda)}(U) \cdot init$.
>

 @kaashif interprets our choices as meant for a physicist audience:
 > - The audience is a bit unclear - physicists? (…)

We recognize that our nomenclature might not be sufficiently clear. Although our work mainly focuses on quantum physics and related fields, we would like the package to be accessible to a wider community, and these changes should help. With this in mind, we suggest the following changes:

1. We can change the terms in the paper to "irreducible representation" and "entry of a matrix." Either option is fine for us, and our main goal is to make the package easier for more people to use.
2. We could also add the standard bi/sesquilinear product notation suggested by @SoongNoonien, along with the physicists' bra-kets. Another option is $(D^{(\lambda)}(U) \text{init}, \text{out})$. In my experience, this notation, which switches the argument order compared to bra-kets, is widely understood.
