### 2026-09-01 #11169 @SoongNoonien
I'd be fine with the suggested changes.
