### 2026-09-02 #11169 @olexandr-konovalov
Many thanks @SoongNoonien and @kaashif for the initial feedback! @davidamaro and @knszm when you will have the next version of the paper, please let us know so that the reviewers can continue with their work. Thanks!
