### 2026-09-02 badab15
Explain single-entry performance advantage
