### 2026-09-07 be2aea7
Changed "matrix element" and "irrep" in the paper.
Sometimes to "matrix entry", "entry of a matrix", or just "entry", whatever sounded best at a particular place. "Irreps" are consistently changed to "irredubile representations"
