+++ paper/paper.md
@@ -14,7 +14,7 @@ authors:
     orcid: "0000-0001-7676-1605"         
     affiliation: 1   
 affiliations:
-  - name: "Research Centre for Quantum Information, Institute of Physics, Slovak Academy of Sciences, Dúbravská cesta 9, Bratislava, Slovakia"
+  - name: "Research Center for Quantum Information, Institute of Physics, Slovak Academy of Sciences, Dúbravská cesta 9, Bratislava, Slovakia"
     index: 1
 date: 29 July 2026
 bibliography: paper.bib
@@ -22,17 +22,17 @@ bibliography: paper.bib
 
 
 # Summary
-[`GroupFunctions.jl`](https://github.com/davidamaro/GroupFunctions.jl)[^author-contributions] is a Julia library for computing individual matrix elements of irreducible representations of $\mathrm{U}(d)$. These matrix elements, called group functions, can be evaluated symbolically or numerically. For $\mathrm{SU}(2)$, they reduce to the Wigner $D$-functions. The library computes these matrix elements in a carrier-space basis enumerated by Gelfand-Tsetlin patterns [@GelfandTsetlin1950]. It can also compute entire representation operators, construct input unitaries from parameterisations common in quantum optics, translate Gelfand-Tsetlin patterns into occupation-number kets, and compute the associated Schur functions.
+[`GroupFunctions.jl`](https://github.com/davidamaro/GroupFunctions.jl)[^author-contributions] is a Julia library for computing individual matrix entries of irreducible representations of $\mathrm{U}(d)$. These entries, called group functions, can be evaluated symbolically or numerically. For $\mathrm{SU}(2)$, they reduce to the Wigner $D$-functions. The library computes these matrix entries in a carrier-space basis enumerated by Gelfand-Tsetlin patterns [@GelfandTsetlin1950]. It can also compute entire representation operators, construct input unitaries from parameterisations common in quantum optics, translate Gelfand-Tsetlin patterns into occupation-number kets, and compute the associated Schur functions.
 
 [^author-contributions]: David Amaro-Alcalá developed the package, its algorithms, and its test suite, and prepared the initial documentation. Konrad Szymański substantially revised and expanded the documentation and developed additional examples, and contributed to the standardisation of the API.
      
 # Statement of need
 
-Representations of the unitary group $\mathrm{U}(d)$ arise in many subfields of physics and mathematics, and computations often reduce to evaluating their matrix elements, called group functions. For an irrep labelled by $\lambda$, the corresponding object is
+Representations of the unitary group $\mathrm{U}(d)$ arise in many subfields of physics and mathematics, and computations often reduce to evaluating their matrix entries, called group functions. For an irreducible representation labelled by $\lambda$, the corresponding object is -- in the mathematicians' and physicists' inner product notation -- the following:
 
-$$ D^{(\lambda)}_{\mathrm{out},\mathrm{init}}(U) = \langle \mathrm{out} \mid D^{(\lambda)}(U) \mid \mathrm{init} \rangle ,$$
+$$ D^{(\lambda)}_{\mathrm{out},\mathrm{init}}(U) = (D^{(\lambda)}(U) \mathrm{init},\mathrm{out}) =\langle \mathrm{out} \mid D^{(\lambda)}(U) \mid \mathrm{init} \rangle ,$$
 
-where $\mathrm{init}$ and $\mathrm{out}$ denote basis states in the representation carrier space.
+where $\mathrm{init}$ and $\mathrm{out}$ denote basis vectors in the representation carrier space.
 
[…]
