### 2026-09-07 1183dea
Clarify representation dimensions and complexity comparison
