+++ paper/paper.md
@@ -40,15 +40,16 @@ In quantum physics, group functions appear in several settings. In quantum optic
 
 Some of these tasks require only a numerical estimate, whereas others require an exact symbolic group function. `GroupFunctions.jl` addresses the latter. Although related software exists (see below), to our knowledge none is designed to compute a single representation-matrix entry as an exact symbolic $D$-function. We therefore compare the cost of computing one such entry with the conventional route that constructs the Lie-algebra representation matrices and exponentiates the resulting full matrix.
 
-Let $\lambda \vdash N$, with $N>1$, and let the compatible basis vectors $A,B$ be supplied directly. For fixed $\lambda,A,B$ and $d\to\infty$, the hook-content formula gives
+The cost of forming the full matrix depends on the dimension $D_\lambda(d)$ of the representation, the number of rows and columns in $D^{(\lambda)}(U)$. Fix a Young diagram $\lambda$ with $N>1$ boxes and at most $d$ rows. The basis vectors are indexed by semistandard Young tableaux of this shape with entries in $\{1,\ldots,d\}$. The hook-content formula counts these tableaux [@Fulton1997, §4.3, p. 55]. Equivalently, the dimension is the Schur polynomial at $d$ arguments equal to 1, because the character at the identity equals the dimension:
 
-$$D_\lambda(d)=\frac{f^{\lambda}}{N!}d^N+O(d^{N-1})=\Theta(d^N).$$
+$$D_\lambda(d)=s_\lambda(\underbrace{1,\ldots,1}_{d})
+=\prod_{(i,j)\in\lambda}\frac{d+j-i}{h_{ij}}.$$
 
-The present implementation evaluates one symbolic $D$-function using $\Theta(d^3)$ structural operations and $\Theta(d^2)$ peak storage. By contrast, even if construction of the $D_\lambda(d)\times D_\lambda(d)$ Lie-algebra matrix is granted at no cost, conventional unstructured scaling-and-squaring/Padé exponentiation with classical dense arithmetic requires $\Theta(D_\lambda(d)^3)=\Theta(d^{3N})$ operations and $\Theta(D_\lambda(d)^2)=\Theta(d^{2N})$ storage. Consequently, the direct-to-exponentiation ratios are
+Here $(i,j)$ denotes a box in row $i$ and column $j$, its content is $j-i$, and its hook length $h_{ij}$ counts the box itself, those to its right in the same row, and those below it in the same column. For fixed $\lambda$, the product has $N$ factors linear in $d$ and a constant denominator, so $D_\lambda(d)=\Theta(d^N)$. For example, $\lambda=(2)$ gives $D_{(2)}(d)=d(d+1)/2$.
 
-$$\Theta\!\left(d^{-3(N-1)}\right) \quad\text{and}\quad \Theta\!\left(d^{-2(N-1)}\right),$$
+To compute one entry directly, we keep the two tableau labels fixed as $d$ grows and supply the corresponding basis vectors, avoiding enumeration of the full basis. In this setting, the current symbolic implementation has time complexity $\Theta(d^3)$ and space complexity $\Theta(d^2)$. The cubic term comes from constructing the tables that index the monomial sum: assembling each length-$d$ row recursively copies successively longer arrays, costing $\Theta(d^2)$ operations per row across $d$ rows. With $N$ fixed, the number of tables and permutations remains bounded as $d$ grows. These estimates use a unit-cost model for array access, copying individual entries, and scalar arithmetic, with space measured by the maximum number of scalar entries stored simultaneously. The dependence on $N$ is absorbed into the constants; bit complexity is excluded.
 
-respectively. Thus, computing one requested $D$-function has an asymptotic advantage over constructing the complete representation matrix in this fixed-$N$, growing-$d$ regime.
[…]
