### 2026-09-08 e404447
Refine paper language and clarify output and complexity notation
