+++ paper/paper.bib
@@ -1,3 +1,11 @@
+@book{BarutRaczka1986,
+  author    = {Barut, Asim Orhan and R{\k{a}}czka, Ryszard},
+  title     = {Theory of Group Representations and Applications},
+  publisher = {PWN-Polish Scientific Publishers},
+  address   = {Warszawa},
+  year      = {1986}
+}
+
 @book{Fulton1997,
   author    = {Fulton, William},
   title     = {Young Tableaux: With Applications to Representation Theory and Geometry},
