+++ paper/paper.md
@@ -22,68 +22,70 @@ bibliography: paper.bib
 
 
 # Summary
-[`GroupFunctions.jl`](https://github.com/davidamaro/GroupFunctions.jl)[^author-contributions] is a Julia library for computing individual matrix entries of irreducible representations of $\mathrm{U}(d)$. These entries, called group functions, can be evaluated symbolically or numerically. For $\mathrm{SU}(2)$, they reduce to the Wigner $D$-functions. The library computes these matrix entries in a carrier-space basis enumerated by Gelfand-Tsetlin patterns [@GelfandTsetlin1950]. It can also compute entire representation operators, construct input unitaries from parameterisations common in quantum optics, translate Gelfand-Tsetlin patterns into occupation-number kets, and compute the associated Schur functions.
+[`GroupFunctions.jl`](https://github.com/davidamaro/GroupFunctions.jl)[^author-contributions] is a Julia library that computes individual matrix entries of irreducible representations of the unitary group $\mathrm{U}(d)$. A representation assigns a matrix to each group element and preserves multiplication. An irreducible representation has no nonzero proper subspace invariant under all its matrices. In a fixed basis, each matrix entry, viewed as a scalar function of the group element, is called a group function or $D$-function. For each requested group function, the user can choose either a symbolic expression or a numerical value at a specified input matrix. For $\mathrm{SU}(2)$, they reduce to the Wigner $D$-functions.
 
-[^author-contributions]: David Amaro-Alcalá developed the package, its algorithms, and its test suite, and prepared the initial documentation. Konrad Szymański substantially revised and expanded the documentation and developed additional examples, and contributed to the standardisation of the API.
+The basis vectors are labelled by Gelfand-Tsetlin patterns [@GelfandTsetlin1950]. The library also computes entire representation matrices, constructs input unitaries from parameterisations common in quantum optics, translates Gelfand-Tsetlin patterns into occupation-number kets, and computes the associated Schur functions.
+
+[^author-contributions]: David Amaro-Alcalá developed the package, its algorithms, and its test suite, and prepared the initial documentation. Konrad Szymański substantially revised and expanded the documentation, developed additional examples, and contributed to the standardisation of the API.
      
 # Statement of need
 
-Representations of the unitary group $\mathrm{U}(d)$ arise in many subfields of physics and mathematics, and computations often reduce to evaluating their matrix entries, called group functions. For an irreducible representation labelled by $\lambda$, the corresponding object is -- in the mathematicians' and physicists' inner product notation -- the following:
+The library serves researchers in mathematics and quantum physics who need symbolic expressions or numerical values for representation matrix entries. For $U\in\mathrm{U}(d)$, let $D^{(\lambda)}(U)$ denote the matrix representing $U$ in an irreducible representation labelled by $\lambda$. The entry with row label $\mathrm{out}$ and column label $\mathrm{init}$ has equivalent expressions in inner-product and bra-ket notation:
 
 $$ D^{(\lambda)}_{\mathrm{out},\mathrm{init}}(U) = (D^{(\lambda)}(U) \mathrm{init},\mathrm{out}) =\langle \mathrm{out} \mid D^{(\lambda)}(U) \mid \mathrm{init} \rangle ,$$
 
-where $\mathrm{init}$ and $\mathrm{out}$ denote basis vectors in the representation carrier space.
+where $\mathrm{init}$ and $\mathrm{out}$ are vectors in the chosen orthonormal basis, and the inner product $(\cdot,\cdot)$ is linear in its first argument.
 
-In mathematics, summing the diagonal group functions gives the trace of the representation matrix of $U$. This trace is the character of the representation and the Schur polynomial of the eigenvalues of $U$, an object central to algebraic combinatorics and symmetric function theory.
[…]
