### 2026-09-08 3f3733e
Distinguish representation dimension from matrix notation
