+++ paper/paper.md
@@ -42,16 +42,16 @@ In quantum optics, group functions give photon transition amplitudes through lin
 
 These tasks can require numerical estimates or exact symbolic group functions. Both are supported by `GroupFunctions.jl`; the following comparison concerns the symbolic case. To our knowledge, none of the related packages discussed below is designed to compute a single representation-matrix entry as an exact symbolic $D$-function. We therefore compare the cost of computing one entry directly with the conventional procedure of constructing the Lie-algebra representation matrices and exponentiating the resulting full matrix.
 
-The cost of forming the full matrix depends on the representation dimension $D_\lambda(d)$, the number of rows and columns in $D^{(\lambda)}(U)$. Fix a Young diagram $\lambda$ with $N>1$ boxes and at most $d$ rows. The basis vectors are indexed by semistandard Young tableaux of this shape with entries in $\{1,\ldots,d\}$. The hook-content formula counts these tableaux [@Fulton1997, §4.3, p. 55]. The dimension $D_\lambda(d)$ of the irreducible representation labelled by $\lambda$ is also the Schur polynomial evaluated at $d$ arguments equal to 1, because a representation's character at the identity equals its dimension:
+The cost of forming the full matrix depends on the representation dimension $r_\lambda(d)$, the number of rows and columns in $D^{(\lambda)}(U)$. Fix a Young diagram $\lambda$ with $N>1$ boxes and at most $d$ rows. The basis vectors are indexed by semistandard Young tableaux of this shape with entries in $\{1,\ldots,d\}$. The hook-content formula counts these tableaux [@Fulton1997, §4.3, p. 55]. The dimension $r_\lambda(d)$ of the irreducible representation labelled by $\lambda$ is also the Schur polynomial evaluated at $d$ arguments equal to 1, because a representation's character at the identity equals its dimension:
 
-$$D_\lambda(d)=s_\lambda(\underbrace{1,\ldots,1}_{d})
+$$r_\lambda(d)=s_\lambda(\underbrace{1,\ldots,1}_{d})
 =\prod_{(i,j)\in\lambda}\frac{d+j-i}{h_{ij}}.$$
 
-The box $(i,j)$ in row $i$ and column $j$ has content $j-i$. Its hook length $h_{ij}$ counts the box itself, those to its right in the same row, and those below it in the same column. For fixed $\lambda$, the product has $N$ factors linear in $d$ and a constant denominator, giving $D_\lambda(d)=\Theta(d^N)$. Here $\Theta(d^k)$ denotes growth bounded above and below by positive constant multiples of $d^k$ as $d\to\infty$. For example, $\lambda=(2)$ gives $D_{(2)}(d)=d(d+1)/2$.
+The box $(i,j)$ in row $i$ and column $j$ has content $j-i$. Its hook length $h_{ij}$ counts the box itself, those to its right in the same row, and those below it in the same column. For fixed $\lambda$, the product has $N$ factors linear in $d$ and a constant denominator, giving $r_\lambda(d)=\Theta(d^N)$. Here $\Theta(d^k)$ denotes growth bounded above and below by positive constant multiples of $d^k$ as $d\to\infty$. For example, $\lambda=(2)$ gives $r_{(2)}(d)=d(d+1)/2$.
 
[…]
