### 2026-09-09 f1d9fc1
Apply paper corrections for JOSS submission
