+++ paper/paper.bib
@@ -1,3 +1,32 @@
+@book{BarutRaczka1986,
+  author    = {Barut, Asim Orhan and R{\k{a}}czka, Ryszard},
+  title     = {Theory of Group Representations and Applications},
+  publisher = {PWN-Polish Scientific Publishers},
+  address   = {Warszawa},
+  year      = {1986}
+}
+
+@book{Fulton1997,
+  author    = {Fulton, William},
+  title     = {Young Tableaux: With Applications to Representation Theory and Geometry},
+  series    = {London Mathematical Society Student Texts},
+  volume    = {35},
+  publisher = {Cambridge University Press},
+  year      = {1997},
+  doi       = {10.1017/CBO9780511626241}
+}
+
+@article{Higham2005,
+  author  = {Higham, Nicholas J.},
+  title   = {The Scaling and Squaring Method for the Matrix Exponential Revisited},
+  journal = {SIAM Journal on Matrix Analysis and Applications},
+  volume  = {26},
+  number  = {4},
+  pages   = {1179--1193},
+  year    = {2005},
+  doi     = {10.1137/04061101X}
+}
+
 @article{GelfandTsetlin1950,
   author  = {Gelfand, I. M. and Tsetlin, M. L.},
   title   = {Finite-dimensional representations of the group of unimodular matrices},
@@ -183,4 +212,3 @@ author = {Daniel Gómez Aguado and Vicent Gimeno and Julio José Moyano-Fernánd
   howpublished = {\url{https://github.com/polyquantique/haarpy}},
   version = {0.0.6}
 }
-
