+++ paper/paper.md
@@ -14,7 +14,7 @@ authors:
     orcid: "0000-0001-7676-1605"         
     affiliation: 1   
 affiliations:
-  - name: "Research Centre for Quantum Information, Institute of Physics, Slovak Academy of Sciences, Dúbravská cesta 9, Bratislava, Slovakia"
+  - name: "Research Center for Quantum Information, Institute of Physics, Slovak Academy of Sciences, Dúbravská cesta 9, Bratislava, Slovakia"
     index: 1
 date: 29 July 2026
 bibliography: paper.bib
@@ -22,57 +22,70 @@ bibliography: paper.bib
 
 
 # Summary
-[`GroupFunctions.jl`](https://github.com/davidamaro/GroupFunctions.jl)[^author-contributions] is a Julia library for computing individual matrix elements of irreducible representations of $\mathrm{U}(d)$. These matrix elements, called group functions, can be evaluated symbolically or numerically. For $\mathrm{SU}(2)$, they reduce to the Wigner $D$-functions. The library computes these matrix elements in a carrier-space basis enumerated by Gelfand-Tsetlin patterns [@GelfandTsetlin1950]. It can also compute entire representation operators, construct input unitaries from parameterisations common in quantum optics, translate Gelfand-Tsetlin patterns into occupation-number kets, and compute the associated Schur functions.
+[`GroupFunctions.jl`](https://github.com/davidamaro/GroupFunctions.jl)[^author-contributions] is a Julia library that computes individual matrix entries of irreducible representations of the unitary group $\mathrm{U}(d)$. A representation assigns a matrix to each group element and preserves multiplication. An irreducible representation has no nonzero proper subspace invariant under all its matrices. In a fixed basis, each matrix entry, viewed as a scalar function of the group element, is called a group function or $D$-function. For each requested group function, the user can choose either a symbolic expression or a numerical value at a specified input matrix. For $\mathrm{SU}(2)$, they reduce to the Wigner $D$-functions.
 
-[^author-contributions]: David Amaro-Alcalá developed the package, its algorithms, and its test suite, and prepared the initial documentation. Konrad Szymański substantially revised and expanded the documentation and developed additional examples, and contributed to the standardisation of the API.
+The basis vectors are labelled by Gelfand-Tsetlin patterns [@GelfandTsetlin1950]. The library also computes entire representation matrices, constructs input unitaries from parameterisations common in quantum optics, translates Gelfand-Tsetlin patterns into occupation-number kets, and computes the associated Schur functions.
+
+[^author-contributions]: David Amaro-Alcalá developed the package, its algorithms, and its test suite, and prepared the initial documentation. Konrad Szymański substantially revised and expanded the documentation, developed additional examples, and contributed to the standardisation of the API.
      
 # Statement of need
 
-Representations of the unitary group $\mathrm{U}(d)$ arise in many subfields of physics and mathematics, and computations often reduce to evaluating their matrix elements, called group functions. For an irrep labelled by $\lambda$, the corresponding object is
+The library serves researchers in mathematics and quantum physics who need symbolic expressions or numerical values for representation matrix entries. For $U\in\mathrm{U}(d)$, let $D^{(\lambda)}(U)$ denote the matrix representing $U$ in an irreducible representation labelled by $\lambda$. The entry with row label $\mathrm{out}$ and column label $\mathrm{init}$ has equivalent expressions in inner-product and bra-ket notation:
 
[…]
