### 2026-09-11 6f00dec
Squashed commit of the following:
commit 54c43d3cef851fd384e893b2c4b833c6bee5723d
Author: Konrad Sz. <k.sz@quantumstat.es>
Date:   Fri Sep 11 21:50:26 2026 +0200

    Terminology change: states->vectors…
    or "basis elements", "state vectors", where appropriate

commit 6abd27f5360fb20cbf2cbafc104abeccf3a2b24f
Author: Konrad Sz. <k.sz@quantumstat.es>
Date:   Fri Sep 11 14:19:47 2026 +0200

    Changed `matrix element` to `(matrix) entry`.

commit 3761a459098c25653adc138d12a7619e9c869758
Author: Konrad Sz. <k.sz@quantumstat.es>
Date:   Fri Sep 11 14:05:48 2026 +0200

    Removed "irreps" from 
[…]
