+++ docs/src/background/group_functions.md
@@ -6,7 +6,7 @@ DocTestSetup = GroupFunctions.doctestsetup()
 
 # Group Functions
 
-The [basis states and GT patterns page](states.md) introduced these labels for basis states, with different irreps corresponding to different exchange symmetries. Here we sketch the mathematical framework behind the library; actual code is shown in the applications pages on [HOM effect](../applications/quantum_optics.md), [qubit transmission with entangled light](../applications/qubit_transmission.md), [sum rules](../applications/sum_rules.md), as well as the related notes on [characters](characters.md) and [immanants](immanants.md).
+The [basis states and GT patterns page](states.md) introduced these labels for basis states, with different irreducible representations corresponding to different exchange symmetries. Here we sketch the mathematical framework behind the library; actual code is shown in the applications pages on [HOM effect](../applications/quantum_optics.md), [qubit transmission with entangled light](../applications/qubit_transmission.md), [sum rules](../applications/sum_rules.md), as well as the related notes on [characters](characters.md) and [immanants](immanants.md).
 
 ## The problem: mode mixing in quantum mechanics
 
@@ -17,7 +17,7 @@ $$|m_1, m_2, \ldots, m_n\rangle = \frac{(a_1^\dagger)^{m_1} \cdots (a_n^\dagger)
 Now suppose the modes get mixed by a unitary transformation $U \in \mathrm{U}(n)$:
 $$a_i^\dagger \mapsto \sum_j U_{ji} a_j^\dagger.$$
 
-What is the output state? Each creation operator in the original Fock state transforms according to the rule above. Expanding the product, we obtain a superposition of Fock states with coefficients that are polynomials in the matrix elements $U_{ij}$.
+What is the output state? Each creation operator in the original Fock state transforms according to the rule above. Expanding the product, we obtain a superposition of Fock states with coefficients that are polynomials in the matrix entries $U_{ij}$.
 
 The function `group_function` computes these coefficients — the transition amplitudes $\langle m' | D^{(\lambda)}(U) | m \rangle$, where $\lambda$ labels the space the operator is acting on (the irreducible representation). The presentation here is slightly more mathematical than a typical quantum optics treatment, for consistency with representation theory literature and to handle more general cases beyond bosons. The partition $\lambda$ labels the symmetry type: $[N,0,0,\ldots]$ for bosons, $[1,1,\ldots,1]$ for fermions, mixed shapes for particles with mixed exchange symmetry. 
 In particular, we keep the label $\lambda$ in the equations to accommodate the more general cases.
@@ -32,7 +32,7 @@ $$D^{(\lambda)}(U)|m\rangle = \frac{1}{\sqrt{\prod_i m_i!}} \prod_{i=1}^{n} \lef
 
 The first step unfolds powers into labeled factors; the second is distributivity. 
 
-So now we have products of operators acting on the vacuum $\ket{0}$; we wish to close it with a bra $\bra{m'}$ to compute the matrix element. Let us act with the creation operators on the vacuum and identify terms proportional to $\ket{m'}$ there; then, the matrix element will be a sum of prefactors leading to $\ket{m'}$.
+So now we have products of operators acting on the vacuum $\ket{0}$; we wish to close it with a bra $\bra{m'}$ to compute the matrix entry. Let us act with the creation operators on the vacuum and identify terms proportional to $\ket{m'}$ there; then, the entry will be a sum of prefactors leading to $\ket{m'}$.
[…]
