+++ docs/src/background/states.md
@@ -4,17 +4,17 @@ CollapsedDocStrings = true
 DocTestSetup = GroupFunctions.doctestsetup()
 ```
 
-# Basis states and Gelfand-Tsetlin patterns
+# Basis vectors and Gelfand-Tsetlin patterns
 
-This library computes matrix elements of U(d) irreducible representations. In quantum optics, for example, these elements describe how passive linear networks transform multiphoton states. Gelfand-Tsetlin (GT) patterns label the basis states between which these matrix elements are evaluated and sometimes give those states a physical interpretation.
+This library computes matrix entries of U(d) irreducible representations. In quantum optics, for example, these entries describe how passive linear networks transform multiphoton states. Gelfand-Tsetlin (GT) patterns label the basis vectors between which these matrix entries are evaluated and sometimes give those vectors a physical interpretation.
 
-Young diagrams commonly identify U(d) irreps. A GT pattern records both the irrep and a state within it. More precisely, its rows record the integer partitions obtained by restricting the irrep along the chain $U(d) \supset U(d-1) \supset \cdots \supset U(1)$. GT patterns thus provide canonical labels for basis vectors, making them useful in subsequent computations.
+Young diagrams commonly identify U(d) irreducible representations. A GT pattern records both the representation and a specific vector within it. More precisely, its rows record the integer partitions obtained by restricting the irreducible representations along the chain $U(d) \supset U(d-1) \supset \cdots \supset U(1)$. GT patterns thus provide canonical labels for basis vectors, making them useful in subsequent computations.
 
 The following standard definitions establish the basic picture.
 
 **Definitions**. Fix a positive integer $d$. The group $U(d)$ consists of the $d\times d$ unitary matrices. A *representation* is a homomorphism $\rho: U(d) \rightarrow GL(V)$ satisfying $\rho(u \cdot u')=\rho(u)\rho(u')$, where $V$ is a finite-dimensional complex vector space and $GL(V)$ is the group of invertible linear maps from $V$ to itself.
 
-A representation is *irreducible* (an *irrep*) if the zero subspace and $V$ itself are its only subspaces invariant under every $\rho(U)$. A *partition* is a nonincreasing integer list $\lambda=[\lambda_1 \ge \cdots \ge \lambda_d]$, or equivalently, a Young diagram with $\lambda_j$ boxes in row $j$. Every finite-dimensional irrep of $U(d)$ corresponds to exactly one partition $\lambda$. We denote the representation by $\rho_\lambda$ and its representation space by $V_\lambda$; the partition becomes the top row of a GT pattern.
+A representation is *irreducible* (an *irrep*) if the zero subspace and $V$ itself are its only subspaces invariant under every $\rho(u)$. A *partition* is a nonincreasing integer list $\lambda=[\lambda_1 \ge \cdots \ge \lambda_d]$, or equivalently, a Young diagram with $\lambda_j$ boxes in row $j$. Every finite-dimensional irreducible representation of $U(d)$ corresponds to exactly one partition $\lambda$. We denote the representation by $\rho_\lambda$ and its representation space by $V_\lambda$; the partition becomes the top row of a GT pattern.
 
 **GT-pattern origin** 
 
@@ -58,30 +58,43 @@ Iterating this interlacing rule along the chain $U(d)\supset U(d-1)\supset\cdots
 
[…]
