### 2026-09-12 34cebaf
Clarify documentation terminology and representation notation
Replace abbreviated irrep terminology with irreducible representation, use basis vectors or elements and matrix entries consistently, and update the navigation headings. Clarify the explanations and examples covering GT patterns, p-weights, Fock states, group functions, immanants, and quantum optics.
