+++ docs/src/tutorial/immanants.md
@@ -6,7 +6,7 @@ DocTestSetup = GroupFunctions.doctestsetup()
 
 # Immanants
 
-Kostant's theorem ([background page](../background/immanants.md)) writes an immanant as a sum of diagonal group functions over the zero-`zweight` states of an irrep. The irrep `λ = [2,1,0]` has two zero-weight states, so the $(2,1)$ immanant of a $3\times3$ matrix is the sum of their two group functions. We pick the states out by their vanishing `zweight`, evaluate the immanant from its definition, and compare:
+Kostant's theorem ([background page](../background/immanants.md)) writes an immanant as a sum of diagonal group functions over the zero-`zweight` basis elements of an irreducible representation. The representation `λ = [2,1,0]` has two zero-weight basis elements, so the $(2,1)$ immanant of a $3\times3$ matrix is the sum of their two group functions. We pick the state vectors out by their vanishing `zweight`, evaluate the immanant from its definition, and compare:
 
 ```@raw html
 <p style="text-align: center;">
@@ -32,7 +32,7 @@ total ≈ imm21(M1)
 
 ## Permanent as special case
 
-The permanent is the immanant of the fully symmetric irrep, $\mathrm{Per}=\mathrm{Imm}^{(n)}$, and there the zero-weight sum collapses to a single term. For three photons in four modes (`λ = [3,0,0,0]`) the group function reproduces the permanent of the relevant submatrix (with a $\sqrt{2}$ normalization factor -- see [the background page](../background/group_functions.md#Bosons:-the-permanent)):
+The permanent is the immanant of the fully symmetric representation, $\mathrm{Per}=\mathrm{Imm}^{(n)}$, and there the zero-weight sum collapses to a single term. For three photons in four modes (`λ = [3,0,0,0]`) the group function reproduces the permanent of the relevant submatrix (with a $\sqrt{2}$ normalization factor -- see [the background page](../background/group_functions.md#Bosons:-the-permanent)):
 
 ```@raw html
 <p style="text-align: center;">
@@ -58,6 +58,6 @@ basis = basis_states([3,0,0,0]);
 state_x = filter(s -> occupation_number(s) == [1,1,1,0], basis)[1]; #modes 1,2,3
 state_y = filter(s -> occupation_number(s) == [0,2,0,1], basis)[1];#modes 2 (twice), 4
 
-M2 = U[[1,2,3], [2,2,4]]; #the modes from above form the indices to extract matrix elements
+M2 = U[[1,2,3], [2,2,4]]; #the modes from above form the indices to extract matrix entries
 group_function([3,0,0,0], state_x, state_y, U) ≈ permanent3(M2) / sqrt(2)
 ```
